Codes Expiry Reminder
Add a function in the product controller with name
send_product_expiry_reminder()-
Module functionality codes expiry reminder :-Need to get notification
email  if product code goi
ng to expire after 30 days .

following Steps need to require to complete the functionality

(1)First we need to calculate the date of expire because in our database
we store the number of days.

(a)To calculate the expire date we need to just add no of days valid to
created date of the product using php .

(2)We need to calculate date after 30 days using php

(a)To calculate the date after 30 days we need to add 30 day to the
current date using php

(3)check if product_expire_date is equal to date after 30 days then we
will store in a array .

(4)and provide the expire date array to the cron job which mail daily to
the admin.